Add PoweredOn/Off abilities to Deployable, execute them in SetPoweredState
