+++ UdpHosts/GameServer/StaticDB/Records/customdata/Todo/SetPoweredStateCommandDef.cs
@@ -3,4 +3,5 @@ namespace GameServer.Data.SDB.Records.customdata;
 public record SetPoweredStateCommandDef : ICommandDef
 {
     public uint Id { get; set; }
+    public bool? PowerOn { get; set; } = null;
 }
