--- UdpHosts/GameServer/Systems/Aptitude/Commands/Other/Todo/SetPoweredStateCommand.cs
@@ -1,19 +1,0 @@
-using GameServer.Data.SDB.Records.customdata;
-
-namespace GameServer.Aptitude;
-
-public class SetPoweredStateCommand : Command, ICommand
-{
-    private SetPoweredStateCommandDef Params;
-
-    public SetPoweredStateCommand(SetPoweredStateCommandDef par)
-: base(par)
-    {
-        Params = par;
-    }
-
-    public bool Execute(Context context)
-    {
-        return true;
-    }
-}
